legacy/requirements.txt
# SilentWitness Python Dependencies
# Install with: pip install -r requirements.txt

# IMPORTANT: tkinter is required for GUI applications
# For embedded Python installations, tkinter-embed provides tkinter support
# For standard Python installations, tkinter is included by default

# Core automation and GUI
pyautogui>=0.9.54
pymsgbox>=0.9.54
pygetwindow>=0.0.9
pytweening>=1.0.7
pyscreeze>=1.3.0
pyrect>=0.2.0

# Essential Python packages (install first)
setuptools>=80.9.0
tkinter-embed>=3.11.0

# Windows-specific functionality
pywin32>=306
pyperclip>=1.8.2
mouse>=0.7.1

# System tray and GUI
pystray>=0.19.5
pillow>=11.3.0
screeninfo>=0.8.1

# Audio and multimedia
pygame>=2.6.1

# Input monitoring
pynput>=1.7.6

# System utilities
psutil>=7.0.0

# Configuration parsing (built-in for Python 3.2+)
# configparser - included in Python standard library
